requirements.txt
flask
flask_login
werkzeug
gunicorn
psycopg2-binary
supabase-py
python-dotenv
